miner: respect maximum_optimized_bitrate; some cleanups
