+++ services/video_scheduler/worker.py
@@ -35,23 +35,14 @@ from video_utils import download_trim_downscale_video, apply_color_space_transfo
 from services.google_drive.google_drive_manager import GoogleDriveManager
 from vidaio_subnet_core import CONFIG
 from vidaio_subnet_core.utilities.storage_client import storage_client
+from vidaio_subnet_core.utilities.tasks import (
+    get_max_optimized_bitrate, get_task_target_resolution,
+)
 
 load_dotenv()
 
 script_dir = os.path.dirname(os.path.abspath(__file__))
 yaml_file_path = os.path.join(script_dir, "pexels_categories.yaml")
-
-
-def get_task_target_resolution(task_type: str) -> Tuple[int, int]:
-    TASK_RESOLUTIONS = {
-        "HD24K": (3840, 2160),
-        "SD2HD": (1920, 1080),
-        "SD24K": (3840, 2160),
-        # "4K28K": (7680, 4320),
-        # "HD28K": (7680, 4320),
-    }
-    return TASK_RESOLUTIONS.get(task_type, (3840, 2160))
-
 
 def clear_queues(redis_conn) -> None:
     """Clear both organic and synthetic queues before starting."""
@@ -443,23 +434,6 @@ async def get_synthetic_requests_paths(num_needed: int, redis_conn: redis.Redis,
             remaining_count -= 1
             
     return uploaded_video_chunks
-
-
-def get_max_optimized_bitrate(task_type: str) -> int:
-    """
-    Get the maximum bitrate of optimized videos (task responses) for a given
-    task type.
-
-    Args:
-        task_type (str): The type of task, e.g. "SD2HD", "HD24K" etc
-    Returns:
-        int: Maximum bitrate of optimized videos in kbit/s
-    """
-    # Bitrate budget: 50 Mbit/s for 4K - necessitates rate control when
-    # encoding H264 videos.
-    width, height = get_task_target_resolution(task_type)
-    n_pixels = width * height
-    return int(50000 * n_pixels / (1920 * 1080))
 
 
 async def main():
